Relocate app and tag data source to data package
