+++ app/src/main/AndroidManifest.xml
@@ -42,12 +42,12 @@
         <activity
             android:name=".ImageActivity_"/>
         <activity
-            android:name=".SettingsActivity_"/>
+            android:name=".settings.SettingsActivity_"/>
         <activity
             android:name=".AmiiboActivity_"
             android:theme="@style/AppTheme2.NoActionBar"/>
         <activity
-            android:name=".TagDataActivity_"/>
+            android:name=".data.TagDataActivity_"/>
     </application>
 
 </manifest>
