+++ app/src/main/java/com/hiddenramblings/tagmo/Actions.java
@@ -1,8 +1,5 @@
 package com.hiddenramblings.tagmo;
 
-/**
- * Created by MAS on 18/06/2016.
- */
 public class Actions {
     public static final String ACTION_EDIT_DATA = "com.hiddenramblings.tagmo.ACTION_EDIT_DATA";
     public static final String ACTION_EDIT_COMPLETE = "com.hiddenramblings.tagmo.ACTION_EDIT_COMPLETE";
