+++ app/src/main/java/com/hiddenramblings/tagmo/AmiiTool.java
@@ -1,8 +1,5 @@
 package com.hiddenramblings.tagmo;
 
-/**
- * Created by MAS on 01/02/2016.
- */
 public class AmiiTool {
     static {
         System.loadLibrary("amiitool");
