+++ app/src/main/java/com/hiddenramblings/tagmo/BrowserAmiibosAdapter.java
@@ -30,6 +30,8 @@ import com.hiddenramblings.tagmo.amiibo.AmiiboSeries;
 import com.hiddenramblings.tagmo.amiibo.AmiiboType;
 import com.hiddenramblings.tagmo.amiibo.Character;
 import com.hiddenramblings.tagmo.amiibo.GameSeries;
+import com.hiddenramblings.tagmo.settings.BrowserSettings;
+import com.hiddenramblings.tagmo.settings.SettingsFragment;
 
 import java.io.File;
 import java.util.ArrayList;
