+++ app/src/main/java/com/hiddenramblings/tagmo/BrowserFoldersAdapter.java
@@ -8,6 +8,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.hiddenramblings.tagmo.settings.BrowserSettings;
+
 import java.io.File;
 import java.util.ArrayList;
