+++ app/src/main/java/com/hiddenramblings/tagmo/HexViewerActivity.java
@@ -3,7 +3,6 @@ package com.hiddenramblings.tagmo;
 import android.annotation.SuppressLint;
 import android.graphics.Color;
 import android.graphics.Typeface;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
