+++ app/src/main/java/com/hiddenramblings/tagmo/data/AppData.java
@@ -1,4 +1,7 @@
-package com.hiddenramblings.tagmo;
+package com.hiddenramblings.tagmo.data;
+
+import com.hiddenramblings.tagmo.R;
+import com.hiddenramblings.tagmo.TagMo;
 
 import java.nio.ByteBuffer;
 
@@ -9,7 +12,7 @@ public class AppData {
 
     public AppData(byte[] appData) throws Exception {
         if (appData.length < APP_FILE_SIZE)
-            throw new Exception("Invalid app data");
+            throw new Exception(TagMo.getStringRes(R.string.invalid_app_data));
 
         this.appData = ByteBuffer.wrap(appData);
     }
