+++ app/src/main/java/com/hiddenramblings/tagmo/data/AppDataInterface.java
@@ -1,4 +1,4 @@
-package com.hiddenramblings.tagmo;
+package com.hiddenramblings.tagmo.data;
 
 interface AppDataInterface {
     void onAppDataChecked(boolean enabled);
