+++ app/src/main/java/com/hiddenramblings/tagmo/data/AppDataTPFragment.java
@@ -1,4 +1,4 @@
-package com.hiddenramblings.tagmo;
+package com.hiddenramblings.tagmo.data;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.widget.Spinner;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.hiddenramblings.tagmo.R;
 
 import org.androidannotations.annotations.AfterViews;
 import org.androidannotations.annotations.EFragment;
@@ -95,10 +97,10 @@ public class AppDataTPFragment extends AppDataFragment {
                 appData.checkLevel(level);
                 txtLevel.setError(null);
             } catch (Exception e) {
-                txtLevel.setError("Must be between 0 and 40");
+                txtLevel.setError(getString(R.string.min_max_error, 0, 40));
             }
         } catch (NumberFormatException e) {
-            txtLevel.setError("Must be between 0 and 40");
+            txtLevel.setError(getString(R.string.min_max_error, 0, 40));
         }
     }
 
@@ -130,10 +132,10 @@ public class AppDataTPFragment extends AppDataFragment {
                 appData.checkHearts(hearts * 4);
                 txtHearts1.setError(null);
             } catch (Exception e) {
-                txtHearts1.setError("Must be between 0 and 20");
+                txtHearts1.setError(getString(R.string.min_max_error, 0, 20));
             }
         } catch (NumberFormatException e) {
-            txtHearts1.setError("Must be between 0 and 20");
+            txtHearts1.setError(getString(R.string.min_max_error, 0, 20));
             txtHearts2.setEnabled(txtHearts1.isEnabled());
         }
     }
