+++ app/src/main/java/com/hiddenramblings/tagmo/data/AppIds.java
@@ -1,4 +1,7 @@
-package com.hiddenramblings.tagmo;
+package com.hiddenramblings.tagmo.data;
+
+import com.hiddenramblings.tagmo.R;
+import com.hiddenramblings.tagmo.TagMo;
 
 import java.util.HashMap;
 
@@ -8,5 +11,6 @@ public class AppIds {
     static {
         appIds.put(AppDataTPFragment.APP_ID, TagMo.getStringRes(R.string.zelda_twilight));
         appIds.put(AppDataSSBFragment.APP_ID, TagMo.getStringRes(R.string.super_smash));
+        appIds.put(-1, TagMo.getStringRes(R.string.unknown));
     }
 }
