+++ app/src/main/java/com/hiddenramblings/tagmo/data/CountryCodes.java
@@ -1,4 +1,4 @@
-package com.hiddenramblings.tagmo;
+package com.hiddenramblings.tagmo.data;
 
 import java.util.HashMap;
