+++ app/src/main/java/com/hiddenramblings/tagmo/settings/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.hiddenramblings.tagmo;
+package com.hiddenramblings.tagmo.settings;
 
 
 import android.annotation.SuppressLint;
@@ -6,6 +6,8 @@ import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.hiddenramblings.tagmo.R;
 
 import org.androidannotations.annotations.EActivity;
