+++ app/src/main/res/layout/fragment_app_data_tp.xml
@@ -1,62 +1,57 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="wrap_content"
+    android:orientation="horizontal"
+    android:baselineAligned="false">
 
     <LinearLayout
-        android:layout_width="match_parent"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:baselineAligned="false">
+        android:layout_weight="1"
+        android:orientation="vertical">
+
+        <TextView
+            style="@style/TagEditor.InputHint"
+            android:text="@string/hearts"/>
 
         <LinearLayout
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:orientation="vertical">
-
-            <TextView
-                style="@style/TagEditor.InputHint"
-                android:text="@string/hearts"/>
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal">
-
-                <EditText
-                    android:id="@+id/txtHearts1"
-                    style="@style/TagEditor.EditNumber"
-                    android:layout_width="0dp"
-                    android:layout_weight="1"
-                    android:importantForAutofill="no"
-                    tools:ignore="LabelFor" />
-
-                <Spinner
-                    android:id="@+id/txtHearts2"
-                    style="@style/TagEditor.Spinner"
-                    android:layout_width="0dp"
-                    android:layout_weight="1"/>
-            </LinearLayout>
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:orientation="vertical">
-
-            <TextView
-                style="@style/TagEditor.InputHint"
-                android:text="@string/level"/>
+            android:orientation="horizontal">
 
             <EditText
-                android:id="@+id/txtLevel"
+                android:id="@+id/txtHearts1"
                 style="@style/TagEditor.EditNumber"
+                android:layout_width="0dp"
+                android:layout_weight="1"
                 android:importantForAutofill="no"
                 tools:ignore="LabelFor" />
+
+            <Spinner
+                android:id="@+id/txtHearts2"
+                style="@style/TagEditor.Spinner"
+                android:layout_width="0dp"
+                android:layout_weight="1"/>
         </LinearLayout>
     </LinearLayout>
-</FrameLayout>
+
+    <LinearLayout
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:orientation="vertical">
+
+        <TextView
+            style="@style/TagEditor.InputHint"
+            android:text="@string/level"/>
+
+        <EditText
+            android:id="@+id/txtLevel"
+            style="@style/TagEditor.EditNumber"
+            android:importantForAutofill="no"
+            tools:ignore="LabelFor" />
+    </LinearLayout>
+</LinearLayout>
